requirements.txt
beautifulsoup4
tqdm
rich
llama-cpp-python
